feat:be double change

See merge request s08-ai-image-sub2/S08P22D102!58

+++ BE/d102/src/main/java/com/ssafy/d102/data/dto/OrganizationDto.java
@@ -13,8 +13,8 @@ public class OrganizationDto {
     private String organizationId;
     private String organizationPw;
     private String organizationName;
-    private Float organizationLng;
-    private Float organizationLat;
+    private Double organizationLng;
+    private Double organizationLat;
     private String organizationSessionId;
     private String organizationRegistDate;
     private String organizationUpdateDate;
